Several fixes caused by missing testing.
